Master

+++ flinkx-ftp/flinkx-ftp-core/src/main/java/com/dtstack/flinkx/ftp/SFtpHandler.java
@@ -24,6 +24,7 @@ import com.jcraft.jsch.JSchException;
 import com.jcraft.jsch.Session;
 import com.jcraft.jsch.SftpATTRS;
 import com.jcraft.jsch.SftpException;
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
@@ -41,31 +42,42 @@ import java.util.Vector;
  * Company: www.dtstack.com
  * @author huyifan.zju@163.com
  */
-public class SFtpHandler implements FtpHandler {
+public class SFtpHandler implements IFtpHandler {
+
     private static final Logger LOG = LoggerFactory.getLogger(SFtpHandler.class);
-    Session session = null;
-    ChannelSftp channelSftp = null;
+
+    private Session session = null;
+
+    private ChannelSftp channelSftp = null;
+
+    private static final String DOT = ".";
+
+    private static final String DOT_DOT = "..";
+
+    private static final String SP = "/";
+
+    private static final String SRC_MAIN = "src/main";
+
+    private static String PATH_NOT_EXIST_ERR = "no such file";
 
     @Override
     public void loginFtpServer(String host, String username, String password, int port, int timeout, String connectMode) {
-        JSch jsch = new JSch(); // 创建JSch对象
+        JSch jsch = new JSch();
         try {
             session = jsch.getSession(username, host, port);
-            // 根据用户名，主机ip，端口获取一个Session对象
-            // 如果服务器连接不上，则抛出异常
             if (session == null) {
                 throw new RuntimeException("login failed. Please check if username and password are correct");
             }
 
-            session.setPassword(password); // 设置密码
+            session.setPassword(password);
             Properties config = new Properties();
             config.put("StrictHostKeyChecking", "no");
-            session.setConfig(config); // 为Session对象设置properties
-            session.setTimeout(timeout); // 设置timeout时间
-            session.connect(); // 通过Session建立链接
+            session.setConfig(config);
+            session.setTimeout(timeout);
+            session.connect();
 
-            channelSftp = (ChannelSftp) session.openChannel("sftp"); // 打开SFTP通道
-            channelSftp.connect(); // 建立SFTP通道的连接
+            channelSftp = (ChannelSftp) session.openChannel("sftp");
+            channelSftp.connect();
         } catch (JSchException e) {
             if(null != e.getCause()){
                 String cause = e.getCause().toString();
@@ -113,10 +125,9 @@ public class SFtpHandler implements FtpHandler {
             SftpATTRS sftpATTRS = channelSftp.lstat(directoryPath);
             return sftpATTRS.isDir();
         } catch (SftpException e) {
-            if (e.getMessage().toLowerCase().equals("no such file")) {
-                String message = String.format("请确认您的配置项path:[%s]存在，且配置的用户有权限读取", directoryPath);
-                LOG.error(message);
-                throw new RuntimeException(message, e);
+            if (e.getMessage().toLowerCase().equals(PATH_NOT_EXIST_ERR)) {
+                LOG.warn("{}", e.getMessage());
+                return false;
             }
             String message = String.format("进入目录：[%s]时发生I/O异常,请确认与ftp服务器的连接正常", directoryPath);
             LOG.error(message);
@@ -133,10 +144,9 @@ public class SFtpHandler implements FtpHandler {
                 isExitFlag = true;
             }
         } catch (SftpException e) {
-            if (e.getMessage().toLowerCase().equals("no such file")) {
-                String message = String.format("请确认您的配置项path:[%s]存在，且配置的用户有权限读取", filePath);
-                LOG.error(message);
-                throw new RuntimeException(message, e);
+            if (e.getMessage().toLowerCase().equals(PATH_NOT_EXIST_ERR)) {
+                LOG.warn("{}", e.getMessage());
+                return false;
             } else {
                 String message = String.format("获取文件：[%s] 属性时发生I/O异常,请确认与ftp服务器的连接正常", filePath);
                 LOG.error(message);
@@ -149,6 +159,7 @@ public class SFtpHandler implements FtpHandler {
     @Override
     public InputStream getInputStream(String filePath) {
         try {
+
             return channelSftp.get(filePath);
         } catch (SftpException e) {
             String message = String.format("读取文件 : [%s] 时出错,请确认文件：[%s]存在且配置的用户有权限读取", filePath, filePath);
@@ -158,24 +169,58 @@ public class SFtpHandler implements FtpHandler {
     }
 
     @Override
+    public List<String> listDirs(String path){
+        if(StringUtils.isBlank(path)) {
+            path = SP;
+        }
+
+        List<String> dirs = new ArrayList<>();
+        if(isDirExist(path)){
+            if(path.equals(DOT) || path.equals(SRC_MAIN)) {
+                return dirs;
+            }
+
+            if(!path.endsWith(SP)) {
+                path = path + SP;
+            }
+
+            try {
+                Vector vector = channelSftp.ls(path);
+                for(int i = 0; i < vector.size(); ++i) {
+                    ChannelSftp.LsEntry le = (ChannelSftp.LsEntry) vector.get(i);
+                    String strName = le.getFilename();
+                    if(!strName.equals(DOT) && !strName.equals(SRC_MAIN) && !strName.equals(DOT_DOT)) {
+                        String filePath = path + strName;
+                        dirs.add(filePath);
+                    }
+                }
+            } catch (SftpException e) {
+                LOG.error("", e);
+            }
+        }
+
+        return dirs;
+    }
+
+    @Override
     public List<String> getFiles(String path) {
         if(StringUtils.isBlank(path)) {
-            path = "/";
+            path = SP;
         }
         List<String> sources = new ArrayList<>();
         if(isDirExist(path)) {
-            if(path.equals(".") || path.equals("src/main")) {
+            if(path.equals(DOT) || path.equals(SRC_MAIN)) {
                 return sources;
             }
-            if(!path.endsWith("/")) {
-                path = path + "/";
+            if(!path.endsWith(SP)) {
+                path = path + SP;
             }
             try {
                 Vector vector = channelSftp.ls(path);
                 for(int i = 0; i < vector.size(); ++i) {
                     ChannelSftp.LsEntry le = (ChannelSftp.LsEntry) vector.get(i);
                     String strName = le.getFilename();
-                    if(!strName.equals(".") && !strName.equals("src/main") && !strName.equals("..")) {
+                    if(!strName.equals(DOT) && !strName.equals(SRC_MAIN) && !strName.equals(DOT_DOT)) {
                         String filePath = path + strName;
                         sources.addAll(getFiles(filePath));
                     }
@@ -200,10 +245,9 @@ public class SFtpHandler implements FtpHandler {
             SftpATTRS sftpATTRS = this.channelSftp.lstat(directoryPath);
             isDirExist = sftpATTRS.isDir();
         } catch (SftpException e) {
-            if (e.getMessage().toLowerCase().equals("no such file")) {
-                LOG.warn(String.format(
-                        "您的配置项path:[%s]不存在，将尝试进行目录创建, errorMessage:%s",
-                        directoryPath, e.getMessage()), e);
+            if (e.getMessage().toLowerCase().equals(PATH_NOT_EXIST_ERR)) {
+                LOG.warn("{}", e.getMessage());
+                LOG.warn("Path [{}] does not exist and will be created.", directoryPath);
                 isDirExist = false;
             }
         }
@@ -231,17 +275,11 @@ public class SFtpHandler implements FtpHandler {
     @Override
     public OutputStream getOutputStream(String filePath) {
         try {
-            this.printWorkingDirectory();
-            String parentDir = filePath.substring(0,
-                    StringUtils.lastIndexOf(filePath, IOUtils.DIR_SEPARATOR));
-            this.channelSftp.cd(parentDir);
-            this.printWorkingDirectory();
-            OutputStream writeOutputStream = this.channelSftp.put(filePath,
-                    ChannelSftp.APPEND);
-            String message = String.format(
-                    "打开FTP文件[%s]获取写出流时出错,请确认文件%s有权限创建，有权限写出等", filePath,
-                    filePath);
+            OutputStream writeOutputStream = this.channelSftp.put(filePath, ChannelSftp.APPEND);
             if (null == writeOutputStream) {
+                String message = String.format(
+                        "打开FTP文件[%s]获取写出流时出错,请确认文件%s有权限创建，有权限写出等", filePath,
+                        filePath);
                 throw new RuntimeException(message);
             }
             return writeOutputStream;
@@ -256,31 +294,37 @@ public class SFtpHandler implements FtpHandler {
 
     private void printWorkingDirectory() {
         try {
-            LOG.info(String.format("current working directory:%s",
-                    this.channelSftp.pwd()));
+            LOG.info("current working directory:{}", this.channelSftp.pwd());
         } catch (Exception e) {
-            LOG.warn(String.format("printWorkingDirectory error:%s",
-                    e.getMessage()));
+            LOG.warn("printWorkingDirectory error:{}", e.getMessage());
         }
     }
 
     @Override
-    public void deleteAllFilesInDir(String dir) {
+    public void deleteAllFilesInDir(String dir, List<String> exclude) {
         if(isDirExist(dir)) {
-            if(!dir.endsWith("/")) {
-                dir = dir + "/";
+            if(!dir.endsWith(SP)) {
+                dir = dir + SP;
             }
+
             try {
                 Vector vector = channelSftp.ls(dir);
                 for(int i = 0; i < vector.size(); ++i) {
                     ChannelSftp.LsEntry le = (ChannelSftp.LsEntry) vector.get(i);
                     String strName = le.getFilename();
-                    if(!strName.equals(".") && !strName.equals("src/main") && !strName.equals("..")) {
+                    if(CollectionUtils.isNotEmpty(exclude) && exclude.contains(strName)){
+                        continue;
+                    }
+
+                    if(!strName.equals(DOT) && !strName.equals(SRC_MAIN) && !strName.equals(DOT_DOT)) {
                         String filePath = dir + strName;
-                        deleteAllFilesInDir(filePath);
+                        deleteAllFilesInDir(filePath, exclude);
                     }
                 }
-                channelSftp.rmdir(dir);
+
+                if(CollectionUtils.isEmpty(exclude)){
+                    channelSftp.rmdir(dir);
+                }
             } catch (SftpException e) {
                 LOG.error("", e);
                 throw new RuntimeException(e);
@@ -302,15 +346,20 @@ public class SFtpHandler implements FtpHandler {
             isDirExist = sftpATTRS.isDir();
         } catch (SftpException e) {
             if(!isDirExist){
-                LOG.info(String.format("正在逐级创建目录 [%s]",directoryPath));
+                LOG.info("Creating a directory step by step [{}]", directoryPath);
                 this.channelSftp.mkdir(directoryPath);
                 return true;
             }
         }
         if(!isDirExist){
-            LOG.info(String.format("正在逐级创建目录 [%s]",directoryPath));
+            LOG.info("Creating a directory step by step [{}]", directoryPath);
             this.channelSftp.mkdir(directoryPath);
         }
         return true;
+    }
+
+    @Override
+    public void rename(String oldPath, String newPath) throws SftpException {
+        channelSftp.rename(oldPath, newPath);
     }
 }
